pyproject.toml
[project]
name = "architectural-styles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "sergeiperel", email = "sergeip.2002@mail.ru" }
]
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "dvc[s3]>=3.64.2",
    "isort>=7.0.0",
    "lightning>=2.6.0",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[project.scripts]
architectural-styles = "architectural_styles.commands:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"


[tool.ruff]
line-length = 100
format = { docstring-code-format = true }
ignore = ["N812"]

[tool.ruff.lint]
select = [
    # isort
    "I",
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # flake8-builtins
    "A",
    # flake8-return
    "RET",
    # flake8-use-pathlib
    "PTH",
    # flynt
    "FLY",
    # mccabe
    "C90",
    # NumPy-specific rules
    "NPY",
    # pep8-naming
    "N",
    # Ruff-specific rules
    "RUF",
    # Pylint - very strict, may be use when have enough time
    "PL",
    # refurb - also may be overwhelming
    "FURB",
]

pycodestyle = { max-doc-length = 105 }

# [tool.ruff.lint.per-file-ignores]
# "__init__.py" = ["E402"]
# "**/{tests,docs,tools}/*" = ["E402"]
# "**/{tests,docs,tools}/*" = ["E402"]
